% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phenos_utils.R
\name{dge_to_vec}
\alias{dge_to_vec}
\title{Helper functions for simulating phenotypes
Convert each dgeMatrix in the list into a vector}
\usage{
dge_to_vec(idx, y)
}
\description{
Helper functions for simulating phenotypes
Convert each dgeMatrix in the list into a vector
}
