% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_main.R
\name{main}
\alias{main}
\title{main function for simulation}
\usage{
main(
  sim_label = 1,
  N_YRI,
  N_CEU,
  N_CHB,
  p_YRI,
  p_CEU,
  p_CHB,
  h2_YRI,
  h2_CEU,
  h2_CHB,
  prare_YRI,
  prare_CEU,
  prare_CHB,
  dist = "gaussian",
  ld = "strong",
  chunk_size,
  cores_used
)
}
\arguments{
\item{sim_label}{denotes the x-th simulation}

\item{N_YRI}{the number of YRI individuals}

\item{N_CEU}{the number of CEU individuals}

\item{N_CHB}{the number of CHB individuals}

\item{p_YRI}{the proportion of causal SNPs in YRI population}

\item{p_CEU}{the proportion of causal SNPs in CEU population}

\item{p_CHB}{the proportion of causal SNPs in CHB population}

\item{h2_YRI}{heritability of this phenotype in YRI population}

\item{h2_CEU}{heritability of this phenotype in CEU population}

\item{h2_CHB}{heritability of this phenotype in CHB population}

\item{prare_YRI}{proportion of causal rare SNPs in YRI population}

\item{prare_CEU}{proportion of causal rare SNPs in CEU population}

\item{prare_CHB}{proportion of causal rare SNPs in CHB population}

\item{dist}{distribution of effect sizes, default is "gaussian".
Options can be "gaussian", "exponential", "laplace".
Note that LDpred simulates non-zero effect sizes from laplace distribution.}

\item{ld}{linkage disequilibrium.
Options can be "strong" or "weak".
We simulate nearby snps together when ld is strong.
We simulate snps evenly spread over all variants when ld is weak.}

\item{chunk_size}{the size of chunk we perform matrix multiplication}

\item{cores_used}{the number of cores used in parallel computing}
}
\value{
phenotypes for three populations and effect sizes for three populations
}
\description{
main function for simulation
}
