% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phenos_utils.R
\name{mul_chunk}
\alias{mul_chunk}
\title{Helper functions for simulating phnotypes
Perform multiplication of a matrix by a vector in chunks}
\usage{
mul_chunk(idx, chunk_size = 100, beta, geno_mat, from, to)
}
\description{
Helper functions for simulating phnotypes
Perform multiplication of a matrix by a vector in chunks
}
