% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phenos_utils.R
\name{mul_parallel}
\alias{mul_parallel}
\title{Helper functions for simulating phnotypes
Use multiple cores to parallely compute large matrix multiplication}
\usage{
mul_parallel(chunk_size = 100, beta, geno_mat, from, to, cores_used)
}
\description{
Helper functions for simulating phnotypes
Use multiple cores to parallely compute large matrix multiplication
}
