% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_var.R
\name{select_var}
\alias{select_var}
\title{select variants}
\usage{
select_var(discovery_geno, discovery_pheno)
}
\arguments{
\item{discovery_geno}{genotype matrix of either YRI or CEU or CHB population}

\item{discovery_pheno}{corresponding phenotypes}
}
\description{
select variants
}
