% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_betas.R
\name{sim_betas}
\alias{sim_betas}
\title{Generate effect sizes betas}
\usage{
sim_betas(M, m, h2, maf_vec, prare)
}
\arguments{
\item{M}{the number of snps}

\item{m}{the number of causal snps}

\item{h2}{heritability}

\item{maf_vec}{the vector of MAF for all snps}

\item{prare}{proportion of rare causal snps}
}
\value{
simulated betas of length M
}
\description{
Generate effect sizes betas
}
