% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phenos_utils.R
\name{sim_E}
\alias{sim_E}
\title{Helper functions for simulating phnotypes
Simulate environmental noise}
\usage{
sim_E(N, h2)
}
\description{
Helper functions for simulating phnotypes
Simulate environmental noise
}
