% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_genos.R
\name{sim_genos}
\alias{sim_genos}
\title{Generate genotypes for three populations YRI, CEU, and CHB.
Note: to run this function, make sure system or terminal or cluster has commands stdpopsim, tskit, and plink2
usually on the PC, make sure these commands have to be under /usr/local/bin}
\usage{
sim_genos(N_YRI, N_CEU, N_CHB, sim_label = 1)
}
\arguments{
\item{N_YRI}{the number of YRI individuals}

\item{N_CEU}{the number of CEU individuals}

\item{N_CHB}{the number of CHB individuals}

\item{sim_label}{denotes the x-th simulation}
}
\value{
output: path for genotypes, a string

freq_stat: path for MAF frequency statistic summary
}
\description{
Generate genotypes for three populations YRI, CEU, and CHB.
Note: to run this function, make sure system or terminal or cluster has commands stdpopsim, tskit, and plink2
usually on the PC, make sure these commands have to be under /usr/local/bin
}
