% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phenos_utils.R
\name{sim_pheno_1pop}
\alias{sim_pheno_1pop}
\title{Helper functions for simulating phnotypes
Simulate phenotypes for one specified population}
\usage{
sim_pheno_1pop(
  pop,
  geno_mat,
  chunk_size,
  cores_used,
  N_YRI,
  N_CEU,
  N_CHB,
  p,
  h2,
  maf_vec,
  prare,
  dist,
  ld
)
}
\description{
Helper functions for simulating phnotypes
Simulate phenotypes for one specified population
}
