% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phenos.R
\name{sim_phenos}
\alias{sim_phenos}
\title{Generate phenotypes for three populations YRI, CEU, and CHB.}
\usage{
sim_phenos(
  geno_mat,
  chunk_size,
  cores_used,
  maf_vec,
  N_YRI,
  N_CEU,
  N_CHB,
  p_YRI,
  p_CEU,
  p_CHB,
  h2_YRI,
  h2_CEU,
  h2_CHB,
  prare_YRI,
  prare_CEU,
  prare_CHB,
  dist,
  ld
)
}
\arguments{
\item{geno_mat}{a BEDMartix of
dimension N (number of individuals from all three populations) by
M (number of SNPs).}

\item{chunk_size}{the size of chunk we perform matrix multiplication}

\item{cores_used}{the number of cores used in parallel computing}

\item{maf_vec}{the vector of MAF for all SNPs}

\item{N_YRI}{the number of YRI individuals}

\item{N_CEU}{the number of CEU individuals}

\item{N_CHB}{the number of CHB individuals}

\item{p_YRI}{the proportion of causal SNPs in YRI population}

\item{p_CEU}{the proportion of causal SNPs in CEU population}

\item{p_CHB}{the proportion of causal SNPs in CHB population}

\item{h2_YRI}{heritability of this phenotype in YRI population}

\item{h2_CEU}{heritability of this phenotype in CEU population}

\item{h2_CHB}{heritability of this phenotype in CHB population}

\item{prare_YRI}{proportion of causal rare SNPs in YRI population}

\item{prare_CEU}{proportion of causal rare SNPs in CEU population}

\item{prare_CHB}{proportion of causal rare SNPs in CHB population}

\item{dist}{distribution of effect sizes, default is "gaussian".
Options can be "gaussian", "exponential", "laplace".
Note that LDpred simulates non-zero effect sizes from laplace distribution.}

\item{ld}{linkage disequilibrium.
Options can be "strong" or "weak".
We simulate nearby snps together when ld is strong.
We simulate snps evenly spread over all variants when ld is weak.}
}
\value{
a list of phenotypes for YRI (pheno_YRI), CEU (pheno_CEU), and CHB (pheno_CHB).
}
\description{
Generate phenotypes for three populations YRI, CEU, and CHB.
}
