% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_betas.R
\name{sim_betas}
\alias{sim_betas}
\title{Generate effect sizes betas}
\usage{
sim_betas(seed = 1234, M, p, h2, maf_vec, prare, dist, ld)
}
\arguments{
\item{seed}{random number seed}

\item{M}{the number of snps}

\item{p}{the proportion of causal snps out of all snps}

\item{h2}{heritability}

\item{maf_vec}{the vector of MAF for all snps}

\item{prare}{proportion of rare causal snps out of causal snps}

\item{dist}{distribution of effect sizes, default is "gaussian".
Options can be "gaussian", "exponential", "laplace".
Note that LDpred simulates non-zero effect sizes from laplace distribution.}

\item{ld}{linkage disequilibrium.
Options can be "strong" or "weak".
We simulate nearby snps together when ld is strong.
We simulate snps evenly spread over all variants when ld is weak.}
}
\value{
simulated betas of length M
}
\description{
Generate effect sizes betas
}
